Improve sdist generation for Python packages
They no longer require torch/featomic/metatensor dependencies
since these are only required to build the wheels. They also
automatically package the Rust/C++ sources in a tarball
