+++ .github/workflows/python-tests.yml
@@ -90,7 +90,7 @@ jobs:
           python -m pip install tox
 
       - name: python build tests
-        run: tox -e build-python
+        run: tox -e build-tests
         env:
           # Use the CPU only version of torch when building/running the code
           PIP_EXTRA_INDEX_URL: https://download.pytorch.org/whl/cpu
