--- MANIFEST.in
@@ -1,27 +1,0 @@
-global-exclude *.pyc
-global-exclude .DS_Store
-
-prune docs
-
-recursive-include featomic *
-recursive-include docs/featomic-json-schema *
-
-# include the minimal crates from the Cargo workspace
-include python/Cargo.toml
-include python/lib.rs
-include featomic-torch/Cargo.toml
-include featomic-torch/lib.rs
-
-include Cargo.*
-include pyproject.toml
-include AUTHORS
-include LICENSE
-
-prune python/tests
-prune python/*.egg-info
-
-prune featomic/tests
-prune featomic/benches/data
-prune featomic/examples/data
-
-exclude tox.ini
