--- python/featomic-torch/.gitignore
@@ -1,1 +1,0 @@
-featomic-torch-*.tar.gz
