+++ python/featomic/AUTHORS
@@ -1,0 +1,1 @@
+../../AUTHORS
