+++ python/featomic/README.rst
@@ -1,0 +1,1 @@
+../../README.rst
