+++ python/featomic/pyproject.toml
@@ -51,27 +51,11 @@ build-backend = "setuptools.build_meta"
 zip-safe = true
 
 [tool.setuptools.packages.find]
-where = ["python/featomic"]
 include = ["featomic*"]
 namespaces = false
 
 ### ======================================================================== ###
 
-
-[tool.ruff.lint]
-select = ["E", "F", "B", "I"]
-ignore = ["B018", "B904"]
-
-[tool.ruff.lint.isort]
-lines-after-imports = 2
-known-first-party = ["featomic"]
-known-third-party = ["torch"]
-
-[tool.ruff.format]
-docstring-code-format = true
-
-### ======================================================================== ###
-
 [tool.pytest.ini_options]
 python_files = ["*.py"]
-testpaths = ["python/featomic/tests"]
+testpaths = ["tests"]
