+++ ruff.toml
@@ -1,0 +1,11 @@
+[lint]
+select = ["E", "F", "B", "I"]
+ignore = ["B018", "B904"]
+
+[lint.isort]
+lines-after-imports = 2
+known-first-party = ["featomic"]
+known-third-party = ["torch"]
+
+[format]
+docstring-code-format = true
