+++ tox.ini
@@ -57,7 +57,7 @@ deps =
     {[testenv]packaging_deps}
 
 commands =
-    pip wheel . {[testenv]build-single-wheel} --wheel-dir {envtmpdir}/dist
+    pip wheel python/featomic {[testenv]build-single-wheel} --wheel-dir {envtmpdir}/dist
 
 
 [testenv:all-deps]
@@ -77,6 +77,7 @@ deps =
     pyscf;platform_system!="Windows"
     wigners
 
+changedir = python/featomic
 commands =
     pytest {[testenv]test_options} {posargs}
 
@@ -87,6 +88,7 @@ deps =
     pytest
     pytest-cov
 
+changedir = python/featomic
 commands =
     pytest {[testenv]test_options} {posargs}
 
@@ -160,36 +162,27 @@ commands =
     ruff check --fix-only {[testenv]lint-folders}
 
 
-[testenv:build-python]
+[testenv:build-tests]
+description = Asserts Pythons package build integrity so one can build sdist and wheels
 package = skip
-# Make sure we can build sdist and a wheel for python
 deps =
-    twine
     build
+    twine  # a tool to check sdist and wheels metadata
+    pip2pi # tool to create PyPI-like package indexes
+    setuptools
 
-allowlist_externals =
-    bash
-
+allowlist_externals = bash
 commands =
     python --version  # print the version of python used in this test
 
-    bash ./scripts/package-featomic-torch.sh python/featomic-torch/
-
-    bash -c "rm -rf {envtmpdir}/dist"
-
-    # check building sdist from a checkout, and wheel from the sdist
-    python -m build . --outdir {envtmpdir}/dist
-
-    # for featomic-torch, we can not build from a sdist until featomic
-    # is available on PyPI, so we build both sdist and wheel from a checkout
-    python -m build python/featomic-torch --sdist --outdir {envtmpdir}/dist
-    python -m build python/featomic-torch --wheel --outdir {envtmpdir}/dist
+    bash ./scripts/build-all-wheels.sh {envtmpdir}
 
     twine check {envtmpdir}/dist/*.tar.gz
     twine check {envtmpdir}/dist/*.whl
 
     # check building wheels directly from the a checkout
-    python -m build . --wheel --outdir {envtmpdir}/dist
+    python -m build python/featomic --wheel --outdir {envtmpdir}/dist
+    python -m build python/featomic-torch --wheel --outdir {envtmpdir}/dist
 
 
 [flake8]
